CR-568 Moved SSO configuration into CFS application
