+++ src/main/java/uk/gov/ons/ctp/integration/censusfieldsvc/CensusFieldSvcApplication.java
@@ -1,5 +1,7 @@
 package uk.gov.ons.ctp.integration.censusfieldsvc;
 
+import com.github.ulisesbocchio.spring.boot.security.saml.configurer.ServiceProviderBuilder;
+import com.github.ulisesbocchio.spring.boot.security.saml.configurer.ServiceProviderConfigurerAdapter;
 import com.godaddy.logging.LoggingConfigs;
 import java.util.HashMap;
 import javax.annotation.PostConstruct;
@@ -13,10 +15,12 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.ImportResource;
 import org.springframework.context.annotation.Primary;
 import org.springframework.http.HttpStatus;
 import org.springframework.integration.annotation.IntegrationComponentScan;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.web.client.RestTemplate;
 import uk.gov.ons.ctp.common.error.RestExceptionHandler;
 import uk.gov.ons.ctp.common.event.EventPublisher;
@@ -133,5 +137,50 @@ public class CensusFieldSvcApplication {
     if (useJsonLogging) {
       LoggingConfigs.setCurrent(LoggingConfigs.getCurrent().useJson());
     }
+  }
+  
+  @Configuration
+  public static class MyServiceProviderConfig extends ServiceProviderConfigurerAdapter {
+
+      @Override
+      public void configure(HttpSecurity http) throws Exception {
+          http.authorizeRequests()
+                  .regexMatchers("/")
+                  .permitAll()
+                  .antMatchers("/completed")
+                  .permitAll()
+                  .antMatchers("/info")
+                  .permitAll()
+                  .antMatchers("/hello3")
+                  .permitAll()
+                  .regexMatchers("/anon/hello")
+                  .permitAll();
+      }
+
+      @Override
+      public void configure(ServiceProviderBuilder serviceProvider) throws Exception {
+          // @formatter:off
+          serviceProvider
+              .metadataGenerator()
+              .entityId("localhost")
+          .and()
+              .sso()
+          .and()
+              .logout()
+              .defaultTargetURL("/afterlogout")
+          .and()
+              .metadataManager()
+              .metadataLocations("classpath:/GoogleIDPMetadata-test.field.census.gov.uk.xml")
+              .defaultIDP("https://accounts.google.com/o/saml2?idpid=C00n4re6c")
+              .refreshCheckInterval(0)
+          .and()
+              .extendedMetadata()
+              .idpDiscoveryEnabled(false)  // disable IDP selection page
+          .and()
+              .keyManager()
+              .privateKeyDERLocation("classpath:/localhost.key.der")
+              .publicKeyPEMLocation("classpath:/localhost.cert");
+          // @formatter:on
+      }
   }
 }
